Replace `apollo-link` package references with local references
