+++ src/__mocks__/mockLinks.ts
@@ -1,12 +1,12 @@
+import Observable from 'zen-observable';
+import { print } from 'graphql/language/printer';
+
 import {
   Operation,
   ApolloLink,
   FetchResult,
-  Observable,
   GraphQLRequest,
-} from 'apollo-link';
-
-import { print } from 'graphql/language/printer';
+} from '../link/core';
 
 interface MockApolloLink extends ApolloLink {
   operation?: Operation;
