+++ src/__tests__/local-state/export.ts
@@ -1,7 +1,8 @@
+import Observable from 'zen-observable';
 import gql from 'graphql-tag';
-import { ApolloLink, Observable } from 'apollo-link';
 import { print } from 'graphql/language/printer';
 
+import { ApolloLink } from '../../link/core';
 import ApolloClient from '../..';
 import { InMemoryCache } from '../../cache/inmemory';
