+++ src/__tests__/local-state/resolvers.ts
@@ -1,8 +1,9 @@
+import Observable from 'zen-observable';
 import gql from 'graphql-tag';
 import { DocumentNode, ExecutionResult } from 'graphql';
 import { assign } from 'lodash';
-import { ApolloLink, Observable } from 'apollo-link';
 
+import { ApolloLink } from '../../link/core';
 import ApolloClient from '../..';
 import mockQueryManager from '../../__mocks__/mockQueryManager';
 import { Observer } from '../../util/Observable';
