+++ src/__tests__/local-state/subscriptions.ts
@@ -1,6 +1,7 @@
+import Observable from 'zen-observable';
 import gql from 'graphql-tag';
-import { ApolloLink, Observable } from 'apollo-link';
 
+import { ApolloLink } from '../../link/core';
 import ApolloClient from '../..';
 import { InMemoryCache } from '../../cache/inmemory';
