+++ src/core/__tests__/QueryManager/links.ts
@@ -1,6 +1,8 @@
 // externals
+import Observable from 'zen-observable';
 import gql from 'graphql-tag';
-import { ApolloLink, Observable } from 'apollo-link';
+
+import { ApolloLink } from '../../../link/core';
 import { InMemoryCache } from '../../../cache/inmemory';
 import { stripSymbols } from '../../../utilities';
