+++ src/react/testing/mocks/types.ts
@@ -1,5 +1,4 @@
-import { ApolloLink, GraphQLRequest, FetchResult } from 'apollo-link';
-
+import { ApolloLink, GraphQLRequest, FetchResult } from '../../../link/core';
 import ApolloClient, { DefaultOptions } from '../../../ApolloClient';
 import { Resolvers } from '../../../core/types';
 import { ApolloCache } from '../../../cache/core/cache';
